refactor(engine): split props-resolver into focused modules
Break the monolithic props-resolver.ts into sibling modules:
- connection-token.ts: connection parsing + handler
- variable-token.ts: variable parsing + handler
- script-evaluator.ts: sandbox/session runner (scriptEvaluator.evaluate,
  initSession) + the sandbox-injected flattenNestedKeys
- split evalStepToken into evalWithPropertyPath / evalWithScript

Move the generic cloneResolvedValue into @activepieces/core-utils
(object-utils) and bump it 0.2.1 -> 0.3.0. props-resolver.ts is left as
pure orchestration.

Behaviour is unchanged, pinned by props-resolver + flow-with-pause +
code-sandbox-session tests. Also records the sandbox-function
serialization, token-name-contract, and per-resolve session-freshness
gotchas in the brain.
