+++ packages/core/utils/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@activepieces/core-utils",
-  "version": "0.2.1",
+  "version": "0.3.0",
   "type": "commonjs",
   "main": "./dist/src/index.js",
   "scripts": {
